+++ src/itemRadio/ItemRadio.tsx
@@ -2,10 +2,10 @@ import React, { Fragment } from 'react'
 import cc from 'classcat'
 
 import { ItemRadioCheckbox } from '../_internals/item/ItemRadioCheckbox.style'
-import { OnChangeParameters } from '../_internals/onChange'
 import { RadioIcon } from '../_internals/radioIcon'
 import { useFocusVisible } from '../_utils/focusVisibleProvider/useFocusVisible'
 import { A11yProps, pickA11yProps } from '../_utils/interfaces'
+import { OnChangeParameters } from '../_utils/onChange'
 import { NormalizeProps } from '../layout/layoutNormalizer'
 import { TextDisplayType } from '../text'